backend/requirements.txt
# Core
fastapi==0.109.2
uvicorn[standard]==0.27.1
python-multipart==0.0.9
pydantic==2.6.1
pydantic-settings==2.1.0

# Google Cloud Vision (Phase 2)
google-cloud-vision==3.5.0

# Enhanced matching
rapidfuzz>=3.0.0
jellyfish>=1.0.0

# LLM (Claude Haiku for OCR normalization)
anthropic>=0.30.0

# Testing
pytest==8.0.0
pytest-asyncio==0.23.4
httpx==0.26.0

# Development
python-dotenv==1.0.1

# Ingestion
pyyaml>=6.0
